requirements.txt
amqp==5.0.9
asgiref==3.5.0
attrs==21.4.0
billiard==3.6.4.0
Brotli==1.0.9
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cssselect2==0.4.1
Django==4.0.1
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-jalali==5.1.0
django-jalali-date==1.0.1
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-taggit==2.1.0
fonttools==4.29.1
html5lib==1.1
idna==3.3
isodate==0.6.1
jdatetime==3.8.2
kombu==5.2.3
lxml==4.7.1
Pillow==9.0.1
platformdirs==2.5.0
prompt-toolkit==3.0.28
pycparser==2.21
pycryptodome==3.14.1
pydyf==0.1.2
pyphen==0.12.0
pytz==2021.3
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
tzdata==2021.5
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
zopfli==0.1.9
